requirements.txt
black==23.1.0
boto3==1.26.95
botocore==1.29.95
click==8.1.3
flake8==6.0.0
isort==5.12.0
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.15
